\name{itfilt-package}
\alias{itfilt-package}
\alias{itfilt}
\docType{package}
\title{
itfilt: A simple package for ion torrent sequences filter.
}
\description{
A simple package for ion torrent sequences filter, including quality control, lenght control, homopolymer remove
}
\details{
\tabular{ll}{
Package: \tab itfilt\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-05-06\cr
License: \tab GPL 3.0\cr
}

}
\author{
Bin Ma
binma01@gmail.com
}
\references{

}

\keyword{ package }

